Split budget table internals
