+++ apps/web/src/ui/tables/budget/table/state/useBudgetTableRangeState.ts
@@ -8,7 +8,7 @@ import {
   getTargetFillMonths,
 } from "@/ui/tables/budget/budgetTableLogic";
 import { fetchBudgetRange } from "@/ui/tables/budget/budgetTableApi";
-import { logBudgetTableError } from "./logBudgetTableError";
+import { logBudgetTableError } from "../logBudgetTableError";
 
 const BATCH_SIZE = 6;
